chore(core): colocate axis style rendering functions
It's cleaner for the axis styling function to live inside the axis module, instead of the styling module assembling functions for the more specific purpose of axes and then exporting those to the axis module.
