+++ source/style.js
@@ -1,20 +1,5 @@
 import { noop } from './helpers.js'
 
-const titleStyles = {
-	titleColor: 'fill',
-	titleFont: 'font-family',
-	titleFontSize: 'font-size',
-	titleFontStyle: 'font-style',
-	titleFontWeight: 'font-weight',
-	titleOpacity: 'opacity'
-}
-
-const tickStyles = {
-	tickColor: 'fill',
-	tickCap: 'stroke-linecap',
-	tickOpacity: 'opacity'
-}
-
 /**
  * read styles from a JavaScript object and apply them as CSS properties
  * @param {object} styles key/value pairs of JavaScript and CSS properties
@@ -35,26 +20,4 @@ const renderStyles = (styles, source) => {
 	}
 }
 
-/**
- * render style instructions for an axis title
- * @param {object} s Vega Lite specification
- * @param {string} channel encoding channel
- * @returns {function(object)} title style rendering function
- */
-const axisTitleStyles = (s, channel) => {
-	const axis = s.encoding[channel].axis
-	return renderStyles(titleStyles, axis)
-}
-
-/**
- * render style instructions for axis ticks
- * @param {object} s Vega Lite specification
- * @param {string} channel encoding channel
- * @returns {function(object)} tick style rendering function
- */
-const axisTickStyles = (s, channel) => {
-	const axis = s.encoding[channel].axis
-	return renderStyles(tickStyles, axis)
-}
-
-export { renderStyles, axisTitleStyles, axisTickStyles }
+export { renderStyles }
